refactor: 페이지 구성 요소 components 폴더 분리 및 메인 파일 레이아웃 적용
